pyproject.toml
[tool.poetry]
name = "pipeline-etl-pdf"
version = "0.1.0"
authors = ["Nelson Paiva <nelsonpaiva10@gmail.com"]
readme = "REAME.md"

[tool.poetry.dependencies]
python = "^3.13.7"
camelot-py = "^1.0.9"
opencv-python = "^4.12.0.88"
matplotlib = "^3.10.7"
ghostscript = "^0.8.1"
psycopg2-binary = "^2.9.11"
sqlalchemy = "^2.0.44"
unidecode = "^1.4.0"
pandas = "^2.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

